Deal with non-i18ned attribute values
Optional i18ned attributes such as 'title' and 'internationalTitle' are in fact not mutually exclusive. GeoServer goes as far as setting 'title' to the first of the i18n values on PUT request if 'title' is not specified but 'internationalTitle' is.
Here we set precedence on the i18n attribute if it is present in the GET response payload and we force a null value on the non-i18ned attribute in the PUT payload to avoid unwanted side effects.